Add files via upload
Uploading initial drop of the Excel file. NOTE: this build works only intermittently because somewhere in the last few years, someone broke the method I was using for getting the data onto the Windows clipboard. There's a solution, but I haven't had a chance to implement it yet. Next upload will (I hope) have that fix in it.